jobmaster/worker/requirements.txt
celery
redis
python-dotenv
psycopg2-binary
# numpy, pandas y scikit-learn comentados - no necesarios para algoritmo simple del enunciado
# numpy
# pandas
# scikit-learn
flower